Issue #69: moving & refactoring tests to the right places
* test_create_and_run_process() -> test_document_collection
* test_get_process() -> test_document_collection
* test_list_document_collection() -> test_project
* adding test.utils
