+++ tests/utils.py
@@ -1,0 +1,26 @@
+#
+# Copyright (c) 2021 Averbis GmbH.
+#
+# This file is part of Averbis Python API.
+# See https://www.averbis.com for further info.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+#
+
+
+def assert_process_equal(expected_process, actual):
+    assert expected_process.name == actual.name
+    assert expected_process.project.name == actual.project.name
+    assert expected_process.pipeline_name == actual.pipeline_name
+    assert expected_process.document_source_name == actual.document_source_name
